Only add new devices when they appear, don't always use the refreshed list.
The issue #6 is that when loading new enumeration (slowly for some
reason on my system) the newly loaded fds will have old events in
them (first opened device will have the oldest one). So if we just
keep using the old devices in that case there will be no gap nor
duplicate events.

This adds a particularly ugly Arc<Mutex<Option<Vec<evdev::Device>>>>
where we must always hold a value inside or we crash. Not sure how to
do this otherwise, but some of the code needs to move the Vec out from
the Mutex, and it doesn't seem it can be done without a container that
can then become empty; but it should never be empty outside the Mutex
lock scope.